wip(daemon): крючки рантайма v5 — черновик
